api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "AI Agent-based diagnostic service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.119.1",
    "logfire>=4.14.1",
    "psycopg2>=2.9.11",
    "pydantic-ai>=1.3.0",
    "pydantic-graph>=0.1.0",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.0.0",
    "redis>=7.0.0",
    "rq>=2.6.0",
    "smolagents[toolkit]>=1.22.0",
    "sqlalchemy>=2.0.44",
    "sqlmodel>=0.0.27",
    "sse-starlette>=3.0.2",
    "supabase>=2.23.3",
    "uvicorn>=0.38.0",
]
